Add player Spawn Fix deltaTime Comment player's movement

+++ includes/script/IconLoading.hpp
@@ -20,7 +20,7 @@ namespace moul
         void start();
         void update();
     private:
-        float m_animTime;
+        double m_animTime;
 
     }; // class Player
 } // namespace moul
